requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
DateTime==5.5
distro==1.9.0
fonttools==4.57.0
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.1
multitasking==0.0.11
newsapi-python==0.2.7
numpy==2.2.5
openai==1.77.0
packaging==25.0
pandas==2.2.3
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yfinance==0.2.56
zope.interface==7.2
